backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyodbc==5.1.0
pymssql==2.2.11
python_dotenv==1.0.1
PyPDF2==3.0.1